requirements.txt
google-api-python-client
streamlit
youtube-transcript-api

